Share strings during processing.
This is a good bet since the same names a reused a lot of time in
practice in big dedukti files.
Searching in the hash table does not seem to be a performance problem
(I tested on files >5Mb).

Here are the output of valgrind and time before and after when running
on a huge file.

==20625== HEAP SUMMARY:
==20625==     in use at exit: 0 bytes in 0 blocks
==20625==   total heap usage: 2,162,617 allocs, 2,162,617 frees, 50,509,407 bytes allocated

==20684== HEAP SUMMARY:
==20684==     in use at exit: 0 bytes in 0 blocks
==20684==   total heap usage: 1,406,573 allocs, 1,406,573 frees, 44,224,883 bytes allocated

real    0m0.854s
user    0m0.427s
sys     0m0.050s

real    0m0.866s
user    0m0.470s
sys     0m0.053s