Add basic module support.
This modification adds full qualification of identifiers in terms, this
allows to avoid conflicts when generating Lua code, this will also
simplify the implementation of the module system.

Each module will be stored in a Lua table, loaded when necessary during
code generation. These tables are made local for the sake of speed.