Move products into the checking phase.
By doing a simple static inspection (see iskind in term.c) we never need
to synthesize the type of a product, hence their typechecking can be
moved in the checking phase.

This allows to clean product type checking (which was buggy), now
the translation of:

    x : A -> B

is

    { tk = tpi, tpi = { term(A), code(A), term(B) } }

which is less confusing than using term boxes for an unclear purpose as
it is done in the Haskell generating version.

These modifications also remove the nesting of type checking functions
in generated terms, which made the semantics of the type checking
routine unclear.