Fix a bug in path generation.
Root elements of the path were screwed because dot patterns were not
taken into account. The test that a variable is applied is now done in
pscope, so we only need to update their usage counter.

Not that variables are marked by np<0, not nd<0, we must be extra
cautious with this! One bug in pat.c related to this issue was fixed.
Changing the type of nd to unsigned could solve this problem...