Add code generation for dot patterns.
The code generator is only slightly changed because the old struct Pat
almost matches with the new one. Though, there is one big difference:
before, we stored array of patterns, now we use array of pointers on
patterns, this makes the copy slightly faster but the access is a bit
longer. I did not benchmark this change.

We need to find a good terminology to speak about the two kinds of
arities (dot patterns v.s. regular patterns). Many functions still use
'ar' or 'arity' when they mean 'regular arity', i.e. the number of
arguments to the constructor which are not dot patterns (i.e. terms).