Add an automatic test system.
I hacked a small test system in Lua, it type checks dedukti files in the
test directory. This is still very hacky because dkparse must be
launched in a specific directory if we want it to work.

I also added a small test ensuring that products cannot be made over
objects.