Handle properly qualified names.
Atoms are still stored as strings however, a function aqual is now
provided in the alloc.c module. This function returns the position of
the first char an identifier's unqualified name. For instance, for the
name "Mod.A.x", aqual returns 6. This function executes in constant time
because its result is provided during the creation of atoms (by the
astrdup function).

This allowed to clean the function gname in the code generator. Indeed,
the module part of a qualified name is now skipped and not mangled.