Handle properly qualified names in patterns.
When a qualified name is encountered in a pattern, currently, we admit
that its module defines it and scope its sub patterns.
Previously we were trying to qualify it with the current module name,
and it failed since a variable name can not contain '.'.

I added a unit test to avoid this kind of bug.