Add support for annotated lambdas.
Lambda terms can now embed the type of their variable. During the type
checking, if a lambda is encountered while in synthesis mode the
annotation is used.

Note that this increases the size of the generated code, translation
writers must take care to generate annotations only when it is mandatory
(on so-called syntactic redexes).